pyproject.toml
[build-system]
requires = ["hatchling >= 1.26"]
build-backend = "hatchling.build"

[project]
name = "physgame-benchmark"
version = "0.0.0"
dependencies = [
    "accelerate==1.6.0",
    "openai==1.70.0",
    "opencv-python==4.11.0.86",
    "pillow==11.1.0",
    "python-dotenv==1.1.0",
    "torch==2.6.0",
    "tqdm==4.67.1",
    "transformers==4.50.3",
]
requires-python = ">=3.12"
